plugins/firefox: fix comparison between signed and unsigned integer expressions

+++ plugins/firefox/trunk/components/EstEIDServiceBase.cpp
@@ -26,7 +26,7 @@ void EstEIDServiceBase::Worker() {
 
 void EstEIDServiceBase::Poll() {
 	try {
-		int nReaders;
+		size_t nReaders;
 
 		{	idAutoLock lock(this);
 			nReaders = mgr.getReaderCount();
